// src/app/models/ApiPerson.ts
import { ApiVertrag } from './ApiVertrag';
import { ApiRolle } from './ApiRolle';
import { ApiRecht } from './ApiRecht';
import { ApiMitarbeiterart } from './ApiMitarbeiterart';
import { ApiFunktion } from './ApiFunktion';
 import { ApiGeschlecht } from './ApiGeschlecht';
import { ApiDienstverwendung } from './ApiDienstverwendung';
import { ApiFreigabeGruppe } from './ApiFreigabeGruppe';
import { ApiOrganisationseinheit } from './ApiOrganisationseinheit';
import { ApiBucher } from './ApiBucher';
import { ApiGetItEntitaet } from './ApiGetItEntitaet';

export interface ApiPerson extends ApiGetItEntitaet {
  titel?: string;
  vorname?: string;
  nachname?: string;
  geschlecht?: ApiGeschlecht;
  gebdat?: string;
  persnr?: string;
  svnr?: string;
  portalUser?: string;
  anmerkung?: string;
  email?: string;
  eintrittsDatum?: string;
  austrittsDatum?: string;
  aktiv?: boolean;
  mobilNummer?: string;
  mobilNummerBmi?: string;
  telefonNummer?: string;
  zimmerNummer?: string;
  emailPrivat?: string;
  strasse?: string;
  hausnr?: string;
  plz?: string;
  ort?: string;
  land?: string;
  staatsangehoerigkeit?: string;
  organisationseinheit?: ApiOrganisationseinheit;
  rolle?: ApiRolle;
  bucher?: ApiBucher;
  recht?: ApiRecht[];
  dienstverwendung?: ApiDienstverwendung;
  mitarbeiterart?: ApiMitarbeiterart;
  firma?: string;
  selbststaendig?: boolean;
  windowsBenutzerkonto?: string;
  leistungskategorie?: string;
  stundensatz?: string;
  geprueft?: boolean;
  stundenkontingentJaehrlich?: string;
  stundenkontingentJaehrlichVertrag?: string;
  leerPdf?: boolean;
  bereitschaftsStundensatz?: string;
  personenverantwortlicher?: ApiPerson;
  teamzuordnung?: string;
  portalDeaktivierung?: number;
  strafregisterbescheid?: string;
  interessenskonflikte?: string;
  teamleiter?: ApiPerson;
  funktion?: ApiFunktion[];
  freigabegruppe?: ApiFreigabeGruppe;
  vertrag?: ApiVertrag[];
  stundenGeplantDiesesJahr?: number;
  stundenGeplantNaechstesJahr?: number;
  stundenGebuchtDiesesJahr?: string;
}